Correct improper usage of WATER_LEVEL
WATER_LEVEL was passed as the number of subdivisions in the water mesh.
Turns out the water level will have to be 0 as long as we use
createGround to create the water mesh (I started looking into
alternatives and decided that switching to something without this
restriction wasn't worth the hassle).
Since WATER_LEVEL has always been 1, this error went unnoticed.