Add special shading for cliffs
Interpolate another texture once the angle between the surface normal
and the xz-plane gets small enough. Getting good u-coordinate for the
texture is difficult; the current version scales atan(norm.z, norm.x) to
get the u-coordinate, but this always looks poor for cliffs that are
linear in the xz plane and can still look bad if the texture scale is
wrong. The texture scale is corrected with the height of the surface,
but it's a mediocre heuristic at best.